feat: add tracking of TMA clicks
