+++ package.json
@@ -65,7 +65,7 @@
     "date-fns": "^3.6.0",
     "dayjs": "^1.11.10",
     "gsap": "^3.12.5",
-    "jsonwebtoken": "^9.0.2",
+    "jsonwebtoken": "8.5.1",
     "jwks-rsa": "^3.1.0",
     "lucide-react": "^0.436.0",
     "mixpanel": "^0.18.0",
